refactor(rpc): 抽取传输无关契约到 traits
- 新增 rpc/traits/，集中 ConsoleCommandService/InstanceService 契约与错误类型
- service/ 精简为纯实现层，调度辅助统一指向 traits
- methods / axum_tests / src-tauri 粘层引用更新到 traits 路径
