+++ server/src/rpc/axum_tests.rs
@@ -12,7 +12,8 @@ use super::*;
 use crate::rpc::{
     methods::server::SendConsoleCommand,
     methods::PERMISSION_SERVER_CONSOLE_SEND,
-    service::{ConsoleCommandService, ConsoleCommandServiceError, InstanceServiceError},
+    service::RpcServices,
+    traits::{ConsoleCommandService, ConsoleCommandServiceError, InstanceServiceError},
     RpcMethodName, RpcPermission,
 };
 
@@ -54,7 +55,7 @@ impl HttpRpcAccessResolver for DenyAll {
 struct NoopInstanceService;
 
 #[async_trait::async_trait]
-impl crate::rpc::service::InstanceService for NoopInstanceService {
+impl crate::rpc::traits::InstanceService for NoopInstanceService {
     async fn list(&self) -> Result<Vec<sealantern_core::instance::Instance>, InstanceServiceError> {
         Ok(Vec::new())
     }
@@ -231,10 +232,7 @@ fn maps_rpc_errors_to_stable_http_statuses() {
 #[tokio::test]
 async fn build_router_dispatches_requests_through_the_public_entry() {
     let svc = Arc::new(RecordingConsoleService { commands: Mutex::new(Vec::new()) });
-    let services = crate::rpc::service::RpcServices::new(
-        svc.clone(),
-        std::sync::Arc::new(NoopInstanceService),
-    );
+    let services = RpcServices::new(svc.clone(), Arc::new(NoopInstanceService));
     let app = crate::rpc::router::build_router(services, AllowConsoleSend);
 
     let response = app
@@ -259,10 +257,7 @@ async fn build_router_dispatches_requests_through_the_public_entry() {
 #[tokio::test]
 async fn build_router_rejects_unprivileged_requests() {
     let svc = Arc::new(RecordingConsoleService { commands: Mutex::new(Vec::new()) });
-    let services = crate::rpc::service::RpcServices::new(
-        svc.clone(),
-        std::sync::Arc::new(NoopInstanceService),
-    );
+    let services = RpcServices::new(svc.clone(), Arc::new(NoopInstanceService));
     let app = crate::rpc::router::build_router(services, DenyAll);
 
     let response = app
@@ -287,10 +282,7 @@ async fn build_router_rejects_unprivileged_requests() {
 #[tokio::test]
 async fn build_router_rejects_invalid_json() {
     let svc = Arc::new(RecordingConsoleService { commands: Mutex::new(Vec::new()) });
-    let services = crate::rpc::service::RpcServices::new(
-        svc.clone(),
-        std::sync::Arc::new(NoopInstanceService),
-    );
+    let services = RpcServices::new(svc.clone(), Arc::new(NoopInstanceService));
     let app = crate::rpc::router::build_router(services, AllowConsoleSend);
 
     let response = app
