+++ server/src/rpc/service/mod.rs
@@ -1,23 +1,12 @@
-//! 供 RPC 方法调用的宿主能力端口。
-
-use std::fmt::Display;
+//! RPC 宿主能力实现层。
+//!
+//! 传输无关契约（trait）定义于 `super::traits`，本目录只承载实现：
+//! [`ServerRuntime`]、[`RpcServices`] 容器与调度辅助函数。
 
 mod console;
-mod instance;
 mod runtime;
 pub(crate) mod services;
 
-pub use console::{dispatch_console_command, ConsoleCommandService, ConsoleCommandServiceError};
-pub use instance::{InstanceService, InstanceServiceError};
+pub use console::dispatch_console_command;
 pub use runtime::ServerRuntime;
 pub use services::RpcServices;
-
-/// 由宿主实现的运行中实例控制台写入能力。
-///
-/// 此 trait 保持应用服务与 Tauri、HTTP 或插件运行时的具体实现解耦。
-pub trait ConsoleCommandExecutor {
-    type Error: Display;
-
-    /// 将命令写入指定实例的控制台。
-    fn send_console_command(&self, instance_id: &str, command: &str) -> Result<(), Self::Error>;
-}
