+++ server/src/rpc/traits/instance.rs
